Moved EMSCEditor to its own file
